# Problem: sender address verification is not decoupled from execution