--- a/app/ante/setup.go
+++ b/app/ante/setup.go
@@ -163,11 +163,6 @@ func (vbd EthValidateBasicDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simu
 			return ctx, errorsmod.Wrapf(errortypes.ErrUnknownRequest, "invalid message type %T, expected %T", msg, (*evmtypes.MsgEthereumTx)(nil))
 		}
 
-		// Validate `From` field
-		if msgEthTx.From != "" {
-			return ctx, errorsmod.Wrapf(errortypes.ErrInvalidRequest, "invalid From %s, expect empty string", msgEthTx.From)
-		}
-
 		txGasLimit += msgEthTx.GetGas()
 
 		txData, err := evmtypes.UnpackTxData(msgEthTx.Data)