--- a/app/ante/utils_test.go
+++ b/app/ante/utils_test.go
@@ -195,7 +195,7 @@ func (s *AnteTestSuite) BuildTestEthTx(
 		input,
 		accesses,
 	)
-	msgEthereumTx.From = from.String()
+	msgEthereumTx.From = from.Bytes()
 	return msgEthereumTx
 }
 
@@ -230,7 +230,6 @@ func (suite *AnteTestSuite) CreateTestTxBuilder(
 	err = msg.Sign(suite.ethSigner, tests.NewSigner(priv))
 	suite.Require().NoError(err)
 
-	msg.From = ""
 	err = builder.SetMsgs(msg)
 	suite.Require().NoError(err)