--- a/rpc/backend/tx_info.go
+++ b/rpc/backend/tx_info.go
@@ -194,7 +194,7 @@ func (b *Backend) GetTransactionReceipt(hash common.Hash) (map[string]interface{
 		return nil, err
 	}
 
-	from, err := ethMsg.GetSender(chainID.ToInt())
+	from, err := ethMsg.GetSenderLegacy(chainID.ToInt())
 	if err != nil {
 		return nil, err
 	}