--- a/rpc/backend/utils.go
+++ b/rpc/backend/utils.go
@@ -104,7 +104,7 @@ func (b *Backend) getAccountNonce(accAddr common.Address, pending bool, height i
 				break
 			}
 
-			sender, err := ethMsg.GetSender(b.chainID)
+			sender, err := ethMsg.GetSenderLegacy(b.chainID)
 			if err != nil {
 				continue
 			}