## Description

WIP: #355

Solution:
- to decouple sender address verification from execution, we must first enforce user setting the From field in the msg
- change `From` field from string to bytes, avoid heavy checksum hex encoding of go-ethereum
- also minimize checksum hex encoding in other places in state machine.

<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
v                               ✰  Thanks for creating a PR! ✰    
v    Before smashing the submit button please review the checkboxes.
v    If a checkbox is n/a - please still include it but + a little note why
☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->

## Benchmark Results

Run the `BenchmarkERC20Transfer` in cronos repo.

```
goos: darwin
goarch: amd64
pkg: github.com/crypto-org-chain/cronos/v2/app
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
                         │ /tmp/after.txt │           /tmp/after2.txt           │
                         │     sec/op     │    sec/op     vs base               │
ERC20Transfer/memdb-12       612.9m ± 19%   602.1m ± 12%        ~ (p=0.818 n=6)
ERC20Transfer/leveldb-12     700.5m ± 18%   593.4m ±  1%  -15.28% (p=0.002 n=6)
ERC20Transfer/memiavl-12     604.1m ± 28%   582.6m ± 23%        ~ (p=0.240 n=6)
geomean                      637.7m         592.6m         -7.07%

                         │ /tmp/after.txt │          /tmp/after2.txt           │
                         │      B/op      │     B/op      vs base              │
ERC20Transfer/memdb-12       176.4Mi ± 0%   174.1Mi ± 0%  -1.32% (p=0.002 n=6)
ERC20Transfer/leveldb-12     178.2Mi ± 0%   175.9Mi ± 0%  -1.30% (p=0.002 n=6)
ERC20Transfer/memiavl-12     171.0Mi ± 0%   168.7Mi ± 0%  -1.36% (p=0.002 n=6)
geomean                      175.2Mi        172.9Mi       -1.33%

                         │ /tmp/after.txt │          /tmp/after2.txt          │
                         │   allocs/op    │  allocs/op   vs base              │
ERC20Transfer/memdb-12        2.802M ± 0%   2.776M ± 0%  -0.93% (p=0.002 n=6)
ERC20Transfer/leveldb-12      2.801M ± 0%   2.775M ± 0%  -0.93% (p=0.002 n=6)
ERC20Transfer/memiavl-12      2.747M ± 0%   2.721M ± 0%  -0.95% (p=0.002 n=6)
geomean                       2.783M        2.757M       -0.93%
```

<!-- Add a description of the changes that this PR introduces and the files that
are the most critical to review.
-->

______

For contributor use:

- [ ] Targeted PR against correct branch (see [CONTRIBUTING.md](https://github.com/cosmos/cosmos-sdk/blob/master/CONTRIBUTING.md#pr-targeting))
- [ ] Linked to Github issue with discussion and accepted design OR link to spec that describes this work.
- [ ] Code follows the [module structure standards](https://github.com/cosmos/cosmos-sdk/blob/master/docs/building-modules/structure.md).
- [ ] Wrote unit and integration [tests](https://github.com/cosmos/cosmos-sdk/blob/master/CONTRIBUTING.md#testing)
- [ ] Updated relevant documentation (`docs/`) or specification (`x/<module>/spec/`)
- [ ] Added  relevant `godoc` [comments](https://blog.golang.org/godoc-documenting-go-code).
- [ ] Added a relevant changelog entry to the `Unreleased` section in `CHANGELOG.md`
- [ ] Re-reviewed `Files changed` in the Github PR explorer

______

For admin use:

- [ ] Added appropriate labels to PR (ex. `WIP`, `R4R`, `docs`, etc)
- [ ] Reviewers assigned
- [ ] Squashed all commits, uses message "Merge pull request #XYZ: [title]" ([coding standards](https://github.com/tendermint/coding/blob/master/README.md#merging-a-pr))